requirements.txt
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
Django==5.2.5
requests==2.31.0
requests-oauthlib==1.3.1
celery==5.3.6
environs==9.5.0
django-environ==0.9.0
django-redis==5.4.0
django-nested-admin==4.0.2
django-ace==1.32.4
django-multiselectfield==0.1.12
django-multiupload==0.6.1
python-slugify==8.0.1
django-silk==5.3.2
django-cors-headers==4.6.0
unidecode==1.3.8
dadata==24.10.0
pillow==11.1.0
django-ckeditor==6.7.2
django-rosetta==0.9.9
django-modeltranslation==0.18.11
django-ipware==5.0.1
marshmallow>=3.13,<4.0
validators==0.28.3
openpyxl==3.1.5
cffi>=1.15.1
psycopg2-binary>=2.9.9
googletrans==4.0.0rc1
xlsxwriter==3.2.5
pandas==2.3.0
daphne==4.1.0
channels_redis==4.2.0
channels==4.0.0
qrcode==8.2